+++ sese/src/native/Windows/security/evp/SHA1Context.cpp
@@ -16,7 +16,7 @@ sese::security::evp::SHA1Context::~SHA1Context() noexcept {
 }
 
 void sese::security::evp::SHA1Context::update(const void *buffer, size_t len) noexcept {
-    CryptHashData(this->hHash, (const BYTE *) buffer, len, 0);
+    CryptHashData(this->hHash, (const BYTE *) buffer, (DWORD) len, 0);
 }
 
 void sese::security::evp::SHA1Context::final() noexcept {
