+++ sese/src/native/Windows/security/evp/SHA384Context.cpp
@@ -18,7 +18,7 @@ sese::security::evp::SHA384Context::~SHA384Context() noexcept {
 }
 
 void sese::security::evp::SHA384Context::update(const void *buffer, size_t len) noexcept {
-    CryptHashData(this->hHash, (const BYTE *) buffer, len, 0);
+    CryptHashData(this->hHash, (const BYTE *) buffer, (DWORD) len, 0);
 }
 
 void sese::security::evp::SHA384Context::final() noexcept {
