+++ sese/src/native/Windows/security/evp/SHA512Context.cpp
@@ -16,7 +16,7 @@ sese::security::evp::SHA512Context::~SHA512Context() noexcept {
 }
 
 void sese::security::evp::SHA512Context::update(const void *buffer, size_t len) noexcept {
-    CryptHashData(this->hHash, (const BYTE *) buffer, len, 0);
+    CryptHashData(this->hHash, (const BYTE *) buffer, (DWORD) len, 0);
 }
 
 void sese::security::evp::SHA512Context::final() noexcept {
